Change type of total material to float

+++ dosing-device/src/main/java/app/web/adapter/persistence/DosingDeviceLastEntity.java
@@ -26,7 +26,7 @@ public class DosingDeviceLastEntity extends DosingSuperEntity {
                                   long version,
                                   String lineName,
                                   int recordNumber,
-                                  int totalMaterial,
+                                  float totalMaterial,
                                   float lastMeasure,
                                   int amountBelowMeasures,
                                   int amountCorrectMeasures,
