+++ weight-module/src/main/java/app/web/weightModule/application/dto/WeightModuleUpdateDto.java
@@ -31,6 +31,6 @@ public class WeightModuleUpdateDto {
         private int amountAboveMeasures;
         private float averageMeasure;
         private int correctMeasuresPercent;
-        private int totalMaterial;
+        private float totalMaterial;
     }
 }
